fix(plugin): limit prio conversion to sched system
The prios in lttng are different from the ones in ftrace, hence a
conversion is needed. However, we must limit this conversion to
well-known trace systems (in this case sched only). We further need to
ensure, we don't mix signed and unsigned data. Hence we split the
converter into two functions.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>